Add files via upload
# Frida Kahlo Exhibition

You've been hired to work on a retrospective of Frida Kahlo's work at a major museum. Your job is to put together the audio tour, but in order to do that you need to create a list of each painting featured in the exhibit, the date it was painted, and its spot in the tour. 

Use your knowledge of Python lists to create a master list of each painting, its date, and its audio tour ID.